refactor: move BSL LS downloader into downloader subpackage, English logs
- Move BslLanguageServerDownloader/BslLanguageServerReleaseChannel to
  com.github._1c_syntax.utils.downloader with @NullMarked package-info
- All log and exception messages in English
