+++ src/test/java/com/github/_1c_syntax/utils/downloader/BslLanguageServerDownloaderTest.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with 1c-syntax utils.
  */
-package com.github._1c_syntax.utils;
+package com.github._1c_syntax.utils.downloader;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledOnOs;
